requirements.txt
beautifulsoup4==4.12.2
playwright==1.40.0
requests==2.31.0
pandas==2.1.4
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
schedule==1.2.1
langchain==0.1.0
openai==1.100.2
scikit-learn==1.3.2
numpy==1.24.4
plotly==5.17.0